Added moderation tests, need to modify code to pass them

+++ back-end/controllers/moderationController.js
@@ -19,15 +19,15 @@ exports.getReports = async (req, res) => {
     let reports;
     try {
         const adminId = req.params.adminId;
-        if (moderationService.isAdmin(adminId)) {
+        const isAdmin = moderationService.isAdmin(adminId);
+        if (isAdmin) {
             reports = await moderationService.getReports();
-            console.log(reports);
             return res.status(200).json({ success: true, reports });
         } else {
             return res.status(403).json({ success: false, message: "Unauthorized access to admin actions" })
         }
     } catch (err) {
-        return res.status(500).json({ success: false, error: "Error while fetching reports: " + err });
+        return res.status(500).json({ success: false, error: "Error getting reports" });
     }
 }
 
@@ -39,14 +39,18 @@ exports.deleteReport = async (req, res) => {
         const adminId = req.params.adminId;
         if (moderationService.isAdmin(adminId)) {
             const reportId = req.params.reportId;
-            await moderationService.deleteReport(reportId);
+            let report = await moderationService.deleteReport(reportId);
 
-            return res.status(200).json({ success: true });
+            if (report == null) {
+                return res.status(400).json({ success: false, error: "Report not found" });
+            }
+
+            return res.status(200).json({ success: true, message: "Report deleted successfully" });
         } else {
             return res.status(403).json({ success: false, message: "Unauthorized access to admin actions" })
         }
     } catch (error) {
-        return res.status(500).json({ success: false, error: "Error while deleting report" + error });
+        return res.status(500).json({ success: false, error: "Error while deleting report"});
     }
 }
 
@@ -60,10 +64,19 @@ exports.banUser = async (req, res) => {
             const userId = req.body.userId;
             const reportId = req.body.reportId;
 
-            await moderationService.ban(userId);
-            await moderationService.deleteReport(reportId);
+            let response1 =  await moderationService.ban(userId);
+
+            if (!response1.success) {
+                return res.status(400).json({ success: false, error: "Invalid user id" });
+            }
+
+            let response2 = await moderationService.deleteReport(reportId);
+
+            if (response2 == null) {
+                return res.status(400).json({ success: false, error: "Report not found" });
+            }
             
-            return res.status(200).json({ success: true });
+            return res.status(200).json({ success: true , message: "User banned successfully"});
         } else {
             return res.status(403).json({ success: false, message: "Unauthorized access to admin actions" })
         }
